feat: r resets current level
Note: need to turn off auto build lighting to work in editor, otherwise
will darken the stage. Needs to be done for each stage individually. There
should not be any problems in normal build, only for editor play. See
https://answers.unity.com/questions/919940/applicationloadlevel0changes-lighting-for-some-rea.html
for instructions.

This commit makes the above changes to the 'tutorial' stage.